EDSC-2057: Adds before_action to prevent errors for users with current sessions
